refactor(wire): move DO-bit patching next to the other wire primitives
forward.rs had grown a second raw-byte parser alongside wire.rs, which
already owns exactly this: scan_ttl_offsets walks the same sections and
recognizes OPT the same way, and patch_id/patch_ttls are the same
patch-in-place shape. Reuse skip_wire_name instead of forking it.

Recognizing OPT by its mandatory root name (RFC 6891 6.1.2) is a 3-byte
peek, which also drops the double name walk the old locate_opt did per
additional record.
